- Extract Reusable Components
